Add test to check for cache clear issue
While the Tag test doesn't test this issue, I think it's a good addition. This particular bug/issue isn't triggered until you ask the library to 'render' the html for you.

The HtmlTest does infact test this bug and when run on current master it WILL fail; however, it passes with the changes suggested.